Fix more memory leaks.
When enumerating lists, a reference to the original pointer needs to be
kept. Otherwise `g_list_free_full` is called on a `NULL` pointer, not
freeing anything. Switch to a proper for loop to fix this.

This fixes the object and interface iterations in `find_hardware` and
`update_device_list`, as well as two menu item iterations in `show_menu`.

The `gtk_widget_destroy` iteration is also simplified to use
`g_list_free_full` with `gtk_widget_destroy` as the callback.

Lastly, remove the unused `sel` variable, which leaks an unused
gtk image.

I think with this all the issues from
https://github.com/raspberrypi-ui/lxplug-bluetooth/issues/15
should be fixed.